@@ -53,9 +53,11 @@ export { default as Loading } from './loading';
 
 export { default as Menu } from './menu';
 
-export { default as Message } from './message';
+export { default as message } from './message';
 
 export { default as Pagination } from './pagination';
+
+export { default as Popover } from './popover';
 
 export { default as Progress } from './progress';
 
@@ -65,6 +67,8 @@ export { default as Rate } from './rate';
 
 export { default as Select } from './select';
 
+export { default as Slider } from './slider';
+
 export { default as Steps } from './steps';
 
 export { default as Switch } from './switch';
@@ -72,6 +76,8 @@ export { default as Switch } from './switch';
 export { default as Table } from './table';
 
 export { default as Tabs } from './tabs';
+
+export { default as Tag } from './tag';
 
 export { default as TimePicker } from './time-picker';
 
