@@ -4,12 +4,17 @@ const Option = Select.Option;
 const OptGroup = Select.OptGroup;
 
 export default class Group extends React.Component {
+  state = {
+    value: 'jack'
+  }
+  onChange = v => {
+    this.setState({
+      value: v
+    })
+  }
   render() {
     return (
-      <Select
-         placeholder='请选择'
-         style={{ width: 200 }}
-      >
+      <Select placeholder='请选择' style={{ width: 200 }} value={this.state.value} onChange={this.onChange}>
          <OptGroup label="Manager">
            <Option value="jack">Jack</Option>
            <Option value="lucy">Lucy</Option>
