@@ -1,6 +1,7 @@
 /**
- * @Author: Zhengfeng.Yao <yzf>
- * @Date:   2017-05-24 15:04:30
- * @Last modified by:   yzf
- * @Last modified time: 2017-05-24 15:04:33
+ * @Author: Yingxi.Hao
+ * @Date:   2017-06-13
+ * @Last modified time: 2017-06-13
  */
+import Tag from './Tag';
+export default Tag;
