@@ -30,8 +30,8 @@ let render = () => {
         <p>连接线</p>
         <Line />
         <p>我是分割线我是分割线</p>
-        <p>过滤</p>
-        <Filter />
+        {/* <p>过滤</p>
+        <Filter /> */}
         <p>我是分割线我是分割线</p>
         <p>拖动示例</p>
         <Drag />
